minor adjustments
* Allows you to specify `container: :none` in the theme so it just produces
content tags outside of a container. Specially, this helps for the cases of
wanting to just format a single line and add to an existing container, like
via Phoenix.LiveView
* changes how the ANSI codes are parsed and chunked so that it only colors
the expected text and not the whole line that color is on
* Adds `\e[39m` to default them which is `default_color` and just specifies
`:text` tag with the intent to use the existing containers default text color
* supports `\e[7m` which is background/forground swap. Attempts to do the swap
if the next token starts with a color. Otherwise ignores it